Finshed refactor of node tool, removed old tool
This commit finishes the refactor of the copyright header tool, allowing
for custom glob strings, as well as an optional --lint flag.
